+++ speech2nutrition/core/domain/entities/food_nutrition_response.py
@@ -10,7 +10,7 @@ class FoodRecord(BaseModel):
     """ The food object """
     score: float
     """ The score of the food record, the higher the score the more likely the food is the correct food """
-    amount: int
+    amount: float
     """ amount in the food's portion unit. if amount is 0, then the the user's unit could not be converted to the food's portion unit"""
     unit_was_transformed: bool
     """ Whether the unit was transformed to the unit given by the food portion unit"""
@@ -21,7 +21,7 @@ class FoodNutritionResponse(BaseModel):
     """ The food record that was found """
     suggestions: List[FoodRecord]
     """ A list of food records that are similar to the food request """
-    user_amount: int
+    user_amount: float
     """ The amount of the food that the user reported """
     user_unit: str
     """ The unit of the amount that the user reported """
